changed all references to money to be an int holding cents rather than a double

+++ src/main/java/com/techelevator/ChewingGum.java
@@ -3,7 +3,7 @@ package com.techelevator;
 public class ChewingGum extends Product {
 
 
-	public ChewingGum(String location, String name, double price, String type, int count) {
+	public ChewingGum(String location, String name, int price, String type, int count) {
 		super(location, name, price, type, count);
 		
 	}
